Added code-logic for user Authentication.

+++ BlogApi/controllers/feed.js
@@ -2,6 +2,7 @@ const { validationResult } = require('express-validator/check');
 
 const fileHelper = require('../helpers/file');
 
+const User = require('../models/user');
 const Post = require('../models/post');
 
 /**
@@ -12,26 +13,27 @@ exports.getPosts = (request, response, next) => {
     const currentPage = request.query.page || 1;
     const perPage = 4;
     let totalItems;
-
-    Post.find().countDocuments().then(count => {
-      totalItems = count;
-      return Post.find()
-        .skip((currentPage - 1) * perPage)
-        .limit(perPage);
-    })
-    .then(posts => {
-      response.status(200).json({
-          message: 'Fetched posts successfully.',
-          posts: posts,
-          totalItems: totalItems
+    Post.find()
+        .countDocuments()
+        .then(count => {
+            totalItems = count;
+            return Post.find()
+                .skip((currentPage - 1) * perPage)
+                .limit(perPage);
+        })
+        .then(posts => {
+            response.status(200).json({
+                message: 'Fetched posts successfully.',
+                posts: posts,
+                totalItems: totalItems
+            });
+        })
+        .catch(err => {
+            if (!err.statusCode) {
+                err.statusCode = 500;
+            }
+            next(err);
         });
-    })
-    .catch(err => {
-      if (!err.statusCode) {
-        err.statusCode = 500;
-      }
-      next(err);
-    });
 };
 
 /**
@@ -78,17 +80,29 @@ exports.createPost = (request, response, next) => {
     const title = request.body.title;
     const content = request.body.content;
 
+    let creator;
+
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: { name: 'Jonas Jsk' }
+        creator: request.userId
     });
-    post.save()
+    post
+        .save()
+        .then(result => {
+            return User.findById(request.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
         .then(result => {
             response.status(201).json({
                 message: 'Post created successfully!',
-                post: result
+                post: post,
+                creator: { _id: creator._id, name: creator.name }
             });
         })
         .catch(err => {
@@ -111,25 +125,27 @@ exports.updatePost = (request, response, next) => {
         error.statusCode = 422;
         throw error;
     }
-
     const title = request.body.title;
     const content = request.body.content;
     let imageUrl = request.body.image;
     if (request.file) {
         imageUrl = request.file.path.replace("\\", "/");
     }
-
     if (!imageUrl) {
         const error = new Error('No file picked.');
         error.statusCode = 422;
         throw error;
     }
-
     Post.findById(postId)
         .then(post => {
             if (!post) {
                 const error = new Error('Could not find post.');
                 error.statusCode = 404;
+                throw error;
+            }
+            if (post.creator.toString() !== request.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
                 throw error;
             }
             if (imageUrl !== post.imageUrl) {
@@ -164,12 +180,23 @@ exports.deletePost = (request, response, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== request.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             // Check logged in user
             fileHelper.clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
-            console.log(result);
+            return User.findById(request.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
             response.status(200).json({ message: 'Deleted post.' });
         })
         .catch(err => {
